refactor: déplace les implems api dans `src/server` (000)
